requirements.txt
streamlit==1.32.2
numpy==1.26.2
pandas==2.2.0
torch==2.2.0
librosa==0.10.1
scipy==1.12.0
whisper==1.1.10
openai-whisper==20231117
yt-dlp==2023.11.16
google-cloud-speech==2.23.0
requests==2.31.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
